refactor: 신청 액션 상태 유틸 위치 이동
