+++ src/features/self-study/ui/SelfStudySection.tsx
@@ -7,10 +7,10 @@ import { TextButton } from "@/shared/ui/Button/TextButton";
 import { NumberButton } from "@/shared/ui/Button/NumberButton";
 import TextField from "@/shared/ui/textField";
 import { dormitoryQueries } from "@/entities/dormitory/api/dormitoryQueries";
+import { createApplicationActionState } from "@/entities/dormitory/lib/applicationActionState";
 import { isStudyApplicationTime } from "@/entities/dormitory/lib/applicationTime";
 import { createStudyPermission } from "@/entities/dormitory/lib/studyPermission";
 import { userQueries } from "@/entities/user/api/userQueries";
-import { createApplicationActionState } from "@/shared/lib/applicationActionState";
 import { getStudyApplyButtonState } from "../lib/getStudyApplyButtonState";
 import { useCurrentTime } from "@/shared/lib/useCurrentTime";
 import { useStudyFilter } from "../model/useStudyFilter";
